feat: Load events from folder
