+++ index.js
@@ -1,6 +1,6 @@
 const { readdirSync } = require('node:fs');
 const { join } = require('node:path');
-const { Client, Collection, Events, GatewayIntentBits, MessageFlags } = require('discord.js');
+const { Client, Collection, GatewayIntentBits } = require('discord.js');
 
 // Configure dotenv
 require('dotenv').config({ quiet: true });
@@ -19,18 +19,7 @@ const clientOptions = {
 // Create a new client instance
 const client = new Client(clientOptions);
 
-// When the client is ready, run this code (only once).
-// The distinction between `client: Client<boolean>` and `readyClient: Client<true>` is important for TypeScript developers.
-// It makes some properties non-nullable.
-client.once(Events.ClientReady, (readyClient) => {
-	console.log(`Ready! Logged in as ${readyClient.user.tag}`);
-});
-
-// Log in to Discord with your client's token
-const token = process.env.DISCORD_JUNO_TOKEN;
-client.login(token);
 client.commands = new Collection();
-
 const foldersPath = join(__dirname, 'commands');
 const commandFolders = readdirSync(foldersPath);
 for (const folder of commandFolders) {
@@ -48,31 +37,21 @@ for (const folder of commandFolders) {
 	}
 }
 
-// listens for the command when it's requested on discord
-client.on(Events.InteractionCreate, async (interaction) => {
-	if (!interaction.isChatInputCommand()) return;
-	// console.log(interaction);
-	const command = interaction.client.commands.get(interaction.commandName);
-	if (!command) {
-		console.error(`No command matching ${interaction.commandName} was found.`);
-		return;
-	}
-	try {
-		await command.execute(interaction);
-	} catch (error) {
-		console.error(error);
-		if (interaction.replied || interaction.deferred) {
-			await interaction.followUp({
-				content: 'There was an error while executing this command!',
-				flags: MessageFlags.Ephemeral,
-			});
-		} else {
-			await interaction.reply({
-				content: 'There was an error while executing this command!',
-				flags: MessageFlags.Ephemeral,
-			});
-		}
-	}
-});
+// Grab all the event files from the events directory
+const eventsPath = join(__dirname, 'events');
+const eventFiles = readdirSync(eventsPath).filter((file) => file.endsWith('.js'));
 
+for (const file of eventFiles) {
+	const filePath = join(eventsPath, file);
+	const event = require(filePath);
+	if (event.once) {
+		client.once(event.name, (...args) => event.execute(...args));
+	} else {
+		client.on(event.name, (...args) => event.execute(...args));
+	}
+}
 
+// Log in to Discord with your client's token
+const token = process.env.DISCORD_JUNO_TOKEN;
+
+client.login(token);
